refactor: reorganize state management and repository structure
