+++ src/lib/data/repositories/LayerRepository.ts
@@ -1,4 +1,4 @@
-import { layerState } from "$lib/presentation/state/LayerState.svelte";
+import { layerState } from "$lib/data/state/LayerState.svelte";
 
 export class LayerRepository {
   static get(): number {
