+++ src/lib/data/repositories/SelectedPanelRepository.ts
@@ -1,5 +1,5 @@
 import type { Panel } from "$lib/domain/entities/Panel";
-import { selectedPanelState } from "$lib/presentation/state/SelectedPanelState.svelte";
+import { selectedPanelState } from "$lib/data/state/SelectedPanelState.svelte";
 
 export class SelectedPanelRepository {
   static get(): Panel | undefined {
