+++ src/lib/data/state/PanelsState.svelte.ts
@@ -1,12 +1,6 @@
-import { PanelsService } from "$lib/data/services/PanelService";
 import { Panel } from "$lib/domain/entities/Panel";
 
 const _panels = $state<Panel[]>([]);
-
-function initialize(layer: number): void {
-  if (_panels.length > 0) return;
-  _panels.splice(0, _panels.length, ...PanelsService.initialize(layer));
-}
 
 function update(panel: Panel): void {
   const idx = _panels.findIndex((x) => x.panelPosition.equals(panel.panelPosition));
@@ -32,7 +26,6 @@ function setAll(panels: Panel[]): void {
 }
 
 export const panelsState = {
-  initialize,
   update,
   remove,
   getAll,
