+++ src/lib/data/state/SelectedPanelState.svelte.ts
