+++ src/routes/+page.svelte
@@ -1,8 +1,8 @@
 <script>
-	import { GameService } from '$lib/domain/services/GameService';
-	import LayeredHexagonPanels from './LayeredHexagonPanels.svelte';
+  import { GameService } from "$lib/services/GameService";
+  import LayeredHexagonPanels from "$lib/presentation/components/LayeredHexagonPanels.svelte";
 
-	GameService.initialize({ layer: 5 });
+  GameService.initialize({ layer: 5 });
 </script>
 
 <LayeredHexagonPanels />
